Refactored file structure and fixed errors in user class and pylintrc
Copied the cosc 310 project files, since they are more up to date than the backend file, then moved them up to the backend directory, and deleted cosc310project. Now there should only be one backend. The duplicate line in user.py has been removed. The pylintrc file was named incorrectly and was not being used by pylint correctly, hence it was also modified slightly. pylint.yml was modified as well, for readability and reduced resource consumption. the myenv folder was also removed, since it was not actually being used, and seemed to be mistakenly added during when the docker files were being created.