Removed venv so it will stop tracking  in github, and added requirements.txt for easy recreation of venv on individual basis
The venv will keep tracking in github regardless of the gitignore til it is deleted from the repo, hence I have removed it. To make reinstalling the packages easy for anyone creating their own venv, the requirements files was created and contains all the packages that were on the original venv. All you have to do it create a new venv, like in the lab, activate it, and run "# Source - https://stackoverflow.com/a
# Posted by Mike Lyons, modified by community. See post 'Timeline' for change history
# Retrieved 2025-11-18, License - CC BY-SA 4.0

pip install -r /path/to/requirements.txt"
(just the pip install part, top part is a citation of where I got the command from"