Removed venv and added the requirements.txt
The venv is set to a specfiic on one of our computers, and hence gives errors at times when trying to access certain files or packages. The solution to this is removing the venv file, and having each dev remake it using the requirements file. To do this, recreate the venv on your machine, then run the command "pip install -r path/path/requirements.txt" to install all the packages. The requirements txt was generated by using pip freeze > requirement.txt on the original venv.